+++ src/Libs/locality/tiling_sequence.c
@@ -152,10 +152,10 @@ static bool loop_flt(loop l )
 static void loop_rwt(loop l, context_p context  ) 
 {  
   contenu_t contenu;
-  int depth;
+  intptr_t depth;
   statement s = loop_body(l);    /* recuperer le coprs de le boucle l */
   contenu = (contenu_t) hash_get(context->contenu, s);
-  depth= (int) hash_get(context->depth, s);   
+  depth= (intptr_t) hash_get(context->depth, s);   
   depth++;
   hash_put(context->depth,stack_head(context->statement_stack), ( void *) depth); 
   hash_put(context->contenu,stack_head(context->statement_stack), (void *) contenu  );
@@ -187,8 +187,8 @@ static bool seq_flt(   )
 static void seq_rwt(sequence sq, context_p context)
 {
   contenu_t contenu;
-  int depth1=0, depth2=0;
-  int max=0;
+  intptr_t depth1=0, depth2=0;
+  intptr_t max=0;
   int i=0;
   list l= sequence_statements(sq);
   contenu=is_a_stencil;
@@ -196,10 +196,10 @@ static void seq_rwt(sequence sq, context_p context)
   MAP(STATEMENT, s,
   {
     contenu = (contenu_t) hash_get(context->contenu, s);
-    if (i==0) depth1 = (int ) hash_get(context->depth, s);
+    if (i==0) depth1 = (intptr_t ) hash_get(context->depth, s);
     if (contenu ==is_a_stencil)  
       {  
-	depth2= (int ) hash_get(context->depth, s);
+	depth2= (intptr_t ) hash_get(context->depth, s);
 	if (depth1!=depth2)
 	  {  
 	    contenu=is_a_no_stencil;
@@ -2568,7 +2568,7 @@ bool tiling_sequence(string module)
      expression_domain, gen_false, gen_null,
      NULL); 
   contenu = (contenu_t) hash_get(context.contenu, stat); 
-  depth = (int ) hash_get(context.depth, stat);
+  depth = (intptr_t ) hash_get(context.depth, stat);
     
   
   if (contenu!=is_a_stencil)
