+++ src/Libs/phrase/comEngine_distribute.c
@@ -108,7 +108,7 @@ static entity  make_new_symbolic_entity(string inName)
   return newEnt;
 }
 
-static void give_value_to_symbolic_entity(entity ent, int val)
+static void give_value_to_symbolic_entity(entity ent, intptr_t val)
 {
   constant constVal = make_constant(is_constant_int, (void *)val);
