+++ src/Libs/transformations/scalarization.c
@@ -617,7 +617,7 @@ static void replace_constant_array_references_walker(reference ref, replace_refe
     {
         /* we know for sure all indices are constant */
         expression offset = reference_offset(ref);
-        int value;
+        intptr_t value;
         if(!expression_integer_value(offset,&value))
             pips_internal_error("reference index should be constants\n");
         /* add one to the value, because 0 seems reserved */
